+++ out/api/types.js
@@ -1,3 +1,7 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.DEVMATE_BACKEND_CAPABILITIES = exports.DEVMATE_BACKEND_PROTOCOL_VERSION = exports.DEVMATE_BACKEND_SERVICE = void 0;
+exports.DEVMATE_BACKEND_SERVICE = 'devmate-backend';
+exports.DEVMATE_BACKEND_PROTOCOL_VERSION = 1;
+exports.DEVMATE_BACKEND_CAPABILITIES = ['chat', 'streaming'];
 //# sourceMappingURL=types.js.map